+++ src/components/InputForm/InsertTextForm.jsx
@@ -2,11 +2,8 @@ import React, { useState } from 'react';
 import styles from './Form.module.scss'
 
 const InsertTextForm = ({formType, guideTxt, type = 'text'}) => {
-
-  // 초기값 세팅 - 아이디, 닉네임, 비밀번호, 비밀번호확인, 이메일, 전화번호, 생년월일
+  // 초기값 세팅 - 아이디, 닉네임, 이메일, 전화번호, 생년월일
   const [id, setId] = useState('');
-  const [password, setPassword] = useState('');
-  const [passwordConfirm, setPasswordConfirm] = useState('');
   const [name, setName] = useState('');
   const [email, setEmail] = useState('');
   const [phone, setPhone] = useState('');
@@ -15,20 +12,15 @@ const InsertTextForm = ({formType, guideTxt, type = 'text'}) => {
   // 오류메세지 상태 저장
   const [idMessage, setIdMessage] = useState('');
   const [nameMessage, setNameMessage] = useState('');
-  const [passwordMessage, setPasswordMessage] = useState('');
-  const [passwordConfirmMessage, setPasswordConfirmMessage] = useState('');
   const [phoneMessage, setPhoneMessage] = useState('');
   const [emailMessage, setEmailMessage] = useState('');
   //  const [birthMessage, setBirthMessage] = useState('');
 
   // 유효성 검사
   const [isId, setIsId] = useState(true)
-  const [isPassword, setIsPassword] = useState(true)
-  const [isPasswordConfirm, setIsPasswordConfirm] = useState(true)
   const [isName, setIsName] = useState(true)
   const [isPhone, setIsPhone] = useState(true)
   const [isEmail, setIsEmail] = useState(true)
-
 
   const onChangeUserID = (e) => {
     const currentId = e.target.value;
@@ -44,34 +36,6 @@ const InsertTextForm = ({formType, guideTxt, type = 'text'}) => {
       setIsId(true)
     }
   };
-
-  const onChangePw = (e) => {
-    const currentPassword = e.target.value;
-    setPassword(currentPassword);
-    const passwordRegExp = /^(?=.*[a-zA-Z])(?=.*[!@#$%^*+=-])(?=.*[0-9]).{8,25}$/;
-    if (!passwordRegExp.test(currentPassword)) {
-      setPasswordMessage(
-        "숫자+영문자+특수문자 조합으로 8자리 이상 입력해주세요!"
-      );
-      setIsPassword(false)
-    } else {
-      setPasswordMessage("안전한 비밀번호 입니다.");
-      setIsPassword(true)
-    }
-  }
-
-  const onChangePwConfirm = (e) => {
-    const currentPasswordConfirm = e.target.value;
-    setPasswordConfirm(currentPasswordConfirm)
-    console.log(currentPasswordConfirm);
-    if ( currentPasswordConfirm !== password) {
-      setPasswordConfirmMessage("떼잉~ 비밀번호가 똑같지 않아요!");
-      setIsPasswordConfirm(false)
-    } else {
-      setPasswordConfirmMessage("똑같은 비밀번호를 입력했습니다.");
-      setIsPasswordConfirm(true)
-    }
-  }
 
   const onChangeUserName = (e) => {
     const currentName = e.target.value
@@ -132,53 +96,23 @@ const InsertTextForm = ({formType, guideTxt, type = 'text'}) => {
       )}
     </>
     }
-    {/* password */}
-    { formType === 'pw' && 
+    {/* userName */}
+    { formType === 'userName' && 
       <>
         <input 
           className={styles.textform} 
           id="input1" 
           type={type} 
           placeholder={guideTxt}
-          onChange={onChangePw}
+          onChange={onChangeUserName}
         /> 
-        { password.length > 0 && (
-          <p className={ `${styles.warningMsg} ${ isPassword ? styles.success : styles.error }` }>{passwordMessage}</p>
+        { name.length > 0 && (
+          <p className={ `${styles.warningMsg} ${ isName ? styles.success : styles.error }` }>{nameMessage}</p>
         )}
       </>
-     }
-      {/* pwConfirm */}
-      { formType === 'pwConfirm' && 
-        <>
-          <input 
-            className={styles.textform} 
-            id="input1" 
-            type={type} 
-            placeholder={guideTxt}
-            onChange={onChangePwConfirm}
-          /> 
-          { passwordConfirm.length > 0 && (
-            <p className={ `${styles.warningMsg} ${ isPasswordConfirm ? styles.success : styles.error }` }>{passwordConfirmMessage}</p>
-          )}
-        </>
-      }
-       {/* userName */}
-       { formType === 'userName' && 
-        <>
-          <input 
-            className={styles.textform} 
-            id="input1" 
-            type={type} 
-            placeholder={guideTxt}
-            onChange={onChangeUserName}
-          /> 
-          { name.length > 0 && (
-            <p className={ `${styles.warningMsg} ${ isName ? styles.success : styles.error }` }>{nameMessage}</p>
-          )}
-        </>
-      }
-       {/* userPhone */}
-       { formType === 'userPhone' && 
+    }
+    {/* userPhone */}
+      { formType === 'userPhone' && 
         <>
           <input 
             className={styles.textform} 
@@ -192,8 +126,8 @@ const InsertTextForm = ({formType, guideTxt, type = 'text'}) => {
           )}
         </>
       }
-        {/* userMail */}
-        { formType === 'userMail' && 
+      {/* userMail */}
+      { formType === 'userMail' && 
         <>
           <input 
             className={styles.textform} 
@@ -207,8 +141,8 @@ const InsertTextForm = ({formType, guideTxt, type = 'text'}) => {
           )}
         </>
       }
-        {/* userAddress */}
-        { formType === 'userAddress' && 
+      {/* userAddress */}
+      { formType === 'userAddress' && 
         <input 
           className={styles.textform} 
           id="input1"
